// src/app/models/contact.ts
// TODO: Add Price

export interface Contact {
  company_name: string;
  email: string;
  icon: string;
  job: string;
  name: string;
  phone: string;
  profile_image: string;
}

// company_name: "Clear Shoot Photography"
// email: "photo@email.com"
// icon: "https://s3.amazonaws.com/honeybook_cdn/assets_users_production/companies/532b267c46ecaccbf1000144/icons/2014-10-02-18-59-41-959.png"
// job: "Photographer & Videographer"
// name: "Luke Peterson"
// phone: "123.456.7890"
// profile_image: "https://s3.amazonaws.com/honeybook_cdn/assets_users_production/users/532b267c46ecaccbf1000143/profile_ph